Add preference open handler
