+++ packages/preferences/src/browser/views/components/preference-node-renderer.ts
@@ -17,7 +17,7 @@
 import { injectable, inject, postConstruct } from '@theia/core/shared/inversify';
 import {
     PreferenceService, ContextMenuRenderer, PreferenceInspection,
-    PreferenceScope, PreferenceProvider, codicon, animationFrame
+    PreferenceScope, PreferenceProvider, codicon, OpenerService, open
 } from '@theia/core/lib/browser';
 import { Preference, PreferenceMenus } from '../../util/preference-types';
 import { PreferenceTreeLabelProvider } from '../../util/preference-tree-label-provider';
@@ -27,7 +27,6 @@ import { JSONValue } from '@theia/core/shared/@phosphor/coreutils';
 import debounce = require('@theia/core/shared/lodash.debounce');
 import { PreferenceTreeModel } from '../../preference-tree-model';
 import { PreferencesSearchbarWidget } from '../preference-searchbar-widget';
-import { WindowService } from '@theia/core/lib/browser/window/window-service';
 import * as markdownit from '@theia/core/shared/markdown-it';
 import * as DOMPurify from '@theia/core/shared/dompurify';
 import URI from '@theia/core/lib/common/uri';
@@ -158,7 +157,7 @@ export abstract class PreferenceLeafNodeRenderer<ValueType extends JSONValue, In
     @inject(PreferencesScopeTabBar) protected readonly scopeTracker: PreferencesScopeTabBar;
     @inject(PreferenceTreeModel) protected readonly model: PreferenceTreeModel;
     @inject(PreferencesSearchbarWidget) protected readonly searchbar: PreferencesSearchbarWidget;
-    @inject(WindowService) protected readonly windowService: WindowService;
+    @inject(OpenerService) protected readonly openerService: OpenerService;
 
     protected headlineWrapper: HTMLDivElement;
     protected gutter: HTMLDivElement;
@@ -213,30 +212,8 @@ export abstract class PreferenceLeafNodeRenderer<ValueType extends JSONValue, In
             // Exclude right click
             if (event.button < 2) {
                 const uri = new URI(event.target.href);
-                if (uri.scheme === 'preference') {
-                    this.selectPreference(uri.path.toString());
-                } else {
-                    // Opens link in external browser
-                    this.windowService.openNewWindow(event.target.href, { external: true });
-                }
+                open(this.openerService, uri);
             }
-        }
-    }
-
-    protected async selectPreference(preferenceId: string): Promise<void> {
-        // Selects the rendered html preference node that does not belong to the commonly used group
-        const selector = `li[data-pref-id="${preferenceId}"]:not([data-node-id^="commonly-used@"])`;
-        const element = document.querySelector(selector);
-        if (element instanceof HTMLElement) {
-            if (element.classList.contains('hidden')) {
-                // We clear the search term as we have clicked on a hidden preference
-                await this.searchbar.updateSearchTerm('');
-                await animationFrame();
-            }
-            element.scrollIntoView({
-                block: 'center'
-            });
-            element.focus();
         }
     }
